+++ pkg/apisvr/deleteserver.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/gin-gonic/gin"
 	"github.com/kimkit/appdaemon/pkg/common"
+	"github.com/kimkit/dbutil"
 )
 
 type DeleteServerController struct {
@@ -13,7 +14,8 @@ type DeleteServerController struct {
 }
 
 func (c *DeleteServerController) POST(ctx *gin.Context) {
-	if err := c.CheckPermission(ctx); err != nil {
+	user, err := c.GetLoginUser(ctx)
+	if err != nil {
 		c.Failure(ctx, err)
 		return
 	}
@@ -27,6 +29,22 @@ func (c *DeleteServerController) POST(ctx *gin.Context) {
 		return
 	}
 
+	rows, err := dbutil.FetchAll(db.Query(
+		"select addr from server where id = ?",
+		id,
+	))
+	if err != nil {
+		c.Failure(ctx, err)
+		return
+	}
+
+	if len(rows) == 0 {
+		c.Failure(ctx, ErrServerNotExist)
+		return
+	}
+
+	addr := rows[0]["addr"]
+
 	_, err = db.Exec(
 		"delete from server where id = ?",
 		id,
@@ -35,6 +53,8 @@ func (c *DeleteServerController) POST(ctx *gin.Context) {
 		c.Failure(ctx, err)
 		return
 	}
+
+	common.Logger.LogInfo("apisvr.DeleteServerController.POST", "server `%s` deleted by `%s`", addr, user)
 
 	c.Success(ctx, nil)
 }
